backend/requirements.txt
fastapi
uvicorn
pydantic
pydantic-settings
python-dotenv
pypdf
sentence-transformers
requests
pytest
httpx
supabase
redis
sse-starlette
python-jose[cryptography]
passlib[bcrypt]
python-multipart
langchain
langchain-core
langchain-openai
langchain-community
